// src/Moryx.Operators.Web/app/src/app/api/models/assignable-operator.ts
/*
 * Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 * Licensed under the Apache License, Version 2.0
*/

/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { IOperatorAssignable } from '../models/i-operator-assignable';
export interface AssignableOperator {
  assignedResources?: Array<IOperatorAssignable> | null;
  firstName?: string | null;
  identifier?: string | null;
  lastName?: string | null;
  pseudonym?: string | null;
  signedIn?: boolean;
}